Marge `FeedRefreshController` with `FeedViewController` since the `UIRefreshControl` creation and configuration now lives in the Storyboard
