--- EssentialFeed/EssentialFeediOS/Feed UI/Controllers/FeedRefreshViewController.swift
@@ -1,31 +1,0 @@
-//
-//  FeedRefreshViewController.swift
-//  EssentialFeediOS
-//
-//  Created by Hitesh on 02/02/22.
-//
-
-import UIKit
-
-protocol FeedRefreshViewControllerDelegate {
-    func didRequestFeedRefresh()
-}
-
-final class FeedRefreshViewController: NSObject, FeedLoadingView {
-    
-    @IBOutlet private var view: UIRefreshControl?
-    
-    var delegate: FeedRefreshViewControllerDelegate?
-        
-    @IBAction func refresh() {
-        delegate?.didRequestFeedRefresh()
-    }
-    
-    func display(_ viewModel: FeedLoadingViewModel) {
-        if viewModel.isLoading {
-            view?.beginRefreshing()
-        } else {
-            view?.endRefreshing()
-        }
-    }
-}
